requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
ccxt==4.5.10
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
cryptography==43.0.1
python-dotenv==1.0.1
websockets==13.1
python-multipart==0.0.9
